Rewrote limelight auto
